feat(load images & refactor FlashCard)
now is able to load all 191 medicine images
refactor FlashCard.vue to use composable
