+++ vuetify-project/public/photo/deck-config.json
@@ -46,7 +46,7 @@
     "鉤藤",
     "雞血藤",
     "麥門冬",
-    "黃岑",
+    "黃芩",
     "黃柏",
     "黃耆",
     "黃蓮"
