+++ vuetify-project/src/components/FlashCard.vue
@@ -1,119 +1,11 @@
 <script setup lang="ts">
-import type { Flashcard } from '@/types/Flashcard'
-import type { Ref } from 'vue'
-import { onMounted, ref } from 'vue'
-type CardDeck = Flashcard[]
+import { useFlashcard } from '@/composables/FlashCard'
+import { onMounted } from 'vue'
 
-const currentDeck: Ref<CardDeck> = ref([])
-const showAnswer = ref(false)
-let cardIndex = 0
-const loadFinished = ref(false)
-const flashcard: Ref<Flashcard> = ref({ imageSrc: '', name: '' })
-/**
- * check if src is valid image(jpeg or png), then store Flashcard to currentDeck, if not valid, throw error
- * @param src : image source
- * @param itemName : item name
- */
-const loadImageBySrc = async (src: string, itemName: string): Promise<void> => {
-  try {
-    const response = await fetch(src)
-    const blob = await response.blob()
-    if (blob.type === 'image/jpeg' || blob.type === 'image/png') {
-      currentDeck.value.push(<Flashcard>{ imageSrc: src, name: itemName })
-    } else {
-      throw new Error('Not a jpeg image')
-    }
-  } catch (error) {
-    console.error(`Error loading image ${src}`, error)
-  }
-}
+const { currentDeck, loadFinished, flashcard, showAnswer, nextCard, initFlashcard } = useFlashcard() // Destructure showAnswer
 
-// const preloadImages = (imageUrls: string[]) => {
-//   imageUrls.forEach((url) => {
-//     const img = new Image()
-//     img.src = url
-//   })
-//   console.log(`Preloading ${imageUrls.length} images...`)
-// }
-
-/**
- * load deck from folder public/photo/itemname*.{png, jpg}
- */
-const LoadDeckfromLocal = async () => {
-  let itemNames: string[] = []
-  const photoFolderPath = `${import.meta.env.BASE_URL}photo/`
-
-  try {
-    const configResponse = await fetch(`${photoFolderPath}deck-config.json`)
-
-    const config: { itemNames: string[] } = await configResponse.json()
-    // console.log(`config: ${JSON.stringify(config)}`)
-
-    if (!Array.isArray(config.itemNames)) {
-      throw new Error('deck-config.json failed to parse itemNames')
-    }
-    itemNames = config.itemNames
-  } catch (error) {
-    console.error('Error loading or parsing deck-config.json', error)
-    itemNames = []
-  }
-
-  // let _ = 0
-  for (const itemName of itemNames) {
-    // if (_++ === 3) break
-    if (!currentDeck.value) {
-      throw new Error('currentDeck is not initialized when loading images')
-    }
-    const imageSrc = `${photoFolderPath}${itemName}.png`
-    let count = 0
-    // get itemName.png
-    try {
-      await loadImageBySrc(imageSrc, itemName)
-      count++
-    } catch (error) {
-      console.error(`Error loading image by src: ${imageSrc}`, error)
-    }
-
-    // get itemNameIndex.png
-    let imageIndex = 1 // Start index from 1 (itemName1.png, itemName2.png, ...)
-
-    while (imageIndex < 4) {
-      const imageSrc = `${photoFolderPath}${itemName}${imageIndex}.png`
-      try {
-        const response = await fetch(imageSrc)
-        const blob = await response.blob()
-        if (blob.type !== 'image/jpeg') {
-          break
-        }
-        currentDeck.value.push(<Flashcard>{ imageSrc: imageSrc, name: itemName })
-        count++
-        imageIndex++
-      } catch (error) {
-        console.error(`Error loading image ${imageSrc}`, error)
-        break
-      }
-    }
-    console.log(`Loaded ${count} images for ${itemName}`)
-  }
-  loadFinished.value = true
-}
-
-/**
- * handle next card button click
- */
-const nextCard = () => {
-  // TODO: randomize the card order
-  cardIndex = (cardIndex + 1) % currentDeck.value.length
-  flashcard.value = currentDeck.value[cardIndex] ?? null
-  showAnswer.value = false
-  console.log(`nextCard: ${flashcard.value.name} cardIndex: ${cardIndex}`)
-  // ...
-}
 onMounted(async () => {
-  await LoadDeckfromLocal()
-  if (currentDeck.value.length > 0) {
-    flashcard.value = currentDeck.value[0]
-  }
+  await initFlashcard()
 })
 </script>
 <template>
